Changed to cope with getDistance() returning a float

+++ trunk/examples/CompassTest/CompassTest.java
@@ -7,7 +7,7 @@ import lejos.nxt.*;
 						
 			while(!Button.ESCAPE.isPressed()) {
 				LCD.clear();
-				LCD.drawInt(compass.getDegrees(), 0, 0);
+				LCD.drawInt((int) compass.getDegrees(), 0, 0);
 				LCD.refresh();
 				Thread.sleep(500);
 			}
